follow @ballob suggestion in https://github.com/home-assistant/architecture/issues/131
