+++ homeassistant/components/utility_meter/services.yaml
@@ -1,13 +1,13 @@
 # Describes the format for available switch services
 
-utility_meter_reset:
+reset:
   description: Resets the counter of an utility meter.
   fields:
     entity_id:
       description: Name(s) of entities to reset 
       example: 'sensor.energy_meter'
 
-utility_meter_start_pause:
+start_pause:
   description: Starts or Pauses the counter of an utility meter.
   fields:
     entity_id:
